start using real GA dashboard for live builds, and move opt out option to somewhere more hidden
